tweak(stories): move ctaSections mockData
