+++ apps/web/src/blocks/CtaSectionsBlock/CtaSectionsBlock.stories.tsx
@@ -1,9 +1,9 @@
 import { faker } from '@faker-js/faker';
+import type { Meta, StoryObj } from '@storybook/react';
 import {
   ctaSection1MockData,
-  ctaSectionTitleMockData
-} from '@mono/web/components/RichText/richTextMockData';
-import type { Meta, StoryObj } from '@storybook/react';
+  ctaSection2MockData
+} from './CtaSectionsMockData';
 
 import type { CtaSectionsBlockType } from '.';
 import CtaSectionsBlock from '.';
@@ -40,7 +40,7 @@ export const Two: Story = {
   args: {
     title: faker.lorem.words(5),
     variant: '2',
-    content: ctaSectionTitleMockData.data
+    content: ctaSection2MockData.data
   }
 };
