+++ apps/web/src/components/RichText/richTextMockData.ts
@@ -1977,110 +1977,4 @@ const richTextMockData: RichTextType = {
   }
 };
 
-export const ctaSection1MockData: RichTextType = {
-  data: {
-    root: {
-      type: 'root',
-      format: '',
-      indent: 0,
-      version: 1,
-      children: [
-        {
-          type: 'paragraph',
-          format: 'center',
-          indent: 0,
-          version: 1,
-          children: [
-            {
-              mode: 'normal',
-              text: 'CTA Section',
-              type: 'eyebrow',
-              style: '',
-              detail: 0,
-              format: 0,
-              version: 1
-            }
-          ],
-          direction: 'ltr',
-          textStyle: '',
-          textFormat: 0
-        },
-        {
-          tag: 'h2',
-          type: 'heading',
-          format: 'center',
-          indent: 0,
-          version: 1,
-          children: [
-            {
-              mode: 'normal',
-              text: 'Action-driving headline that creates urgency',
-              type: 'text',
-              style: '',
-              detail: 0,
-              format: 0,
-              version: 1
-            }
-          ],
-          direction: 'ltr'
-        },
-        {
-          type: 'paragraph',
-          format: 'center',
-          indent: 0,
-          version: 1,
-          children: [
-            {
-              mode: 'normal',
-              text: 'Add one or two compelling sentences that reinforce your main value proposition and overcome final objections. End with a clear reason to act now. Align this copy with your CTA button text.',
-              type: 'text',
-              style: '',
-              detail: 0,
-              format: 0,
-              version: 1
-            }
-          ],
-          direction: 'ltr',
-          textStyle: '',
-          textFormat: 0
-        }
-      ],
-      direction: 'ltr'
-    }
-  }
-};
-
-export const ctaSectionTitleMockData: RichTextType = {
-  data: {
-    root: {
-      type: 'root',
-      format: '',
-      indent: 0,
-      version: 1,
-      children: [
-        {
-          tag: 'h2',
-          type: 'heading',
-          format: 'center',
-          indent: 0,
-          version: 1,
-          children: [
-            {
-              mode: 'normal',
-              text: 'Action-driving headline that creates urgency',
-              type: 'text',
-              style: '',
-              detail: 0,
-              format: 0,
-              version: 1
-            }
-          ],
-          direction: 'ltr'
-        }
-      ],
-      direction: 'ltr'
-    }
-  }
-};
-
 export default richTextMockData;
